requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
archtool==0.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.1.0
centrifuge-python==0.3.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
fastapi==0.115.8
frozenlist==1.5.0
furl==2.1.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
minio==7.2.15
multidict==6.1.0
orderedmultidict==1.0.1
propcache==0.2.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==11.0.3
yarl==1.18.3
